requirements.txt
pyaudio
openai
python-dotenv
SpeechRecognition
vosk
python-kasa
asyncio-extras